refactor/repackeged exceptions and security->config
